Render Tiles
The render now renders tiles in the image. How many tiles make up the
image, and where the tiles go is currently hard coded. The idea here is
to make this easier to multi-thread safely. This is the first step to
doing that, because we can now easily render each tile in isolation.

After that works, it's a matter of tweaking for performance.